Refactor: Split large seed data files into thematic modules
Address Sourcery AI code review feedback by improving code organization:

**Changes Made:**

1. **Fixed ITEM_TYPE duplicate entries**
   - Removed duplicate `EQUIPMENT` type from ItemTypes.ts
   - Changed all `ITEM_TYPE.EQUIPMENT` references to `ITEM_TYPE.EQUIP` (19 occurrences)
   - Ensures consistent item type usage across the codebase

2. **Split Quests.ts (757 lines → 3 modular files)**
   - Created `quests/main-quests.ts` - Main story quest line (20 quests)
   - Created `quests/side-quests.ts` - Optional side content (10 quests)
   - Updated Quests.ts to re-export from modules

3. **Split Skills.ts (471 lines → 4 modular files)**
   - Created `skills/warrior-skills.ts` - Tank/DPS skills (11 skills)
   - Created `skills/mage-skills.ts` - Elemental magic (12 skills)
   - Created `skills/rogue-skills.ts` - Critical/stealth (12 skills)
   - Updated Skills.ts to re-export and combine all trees

4. **Split ExpandedItems.ts (553 lines → 7 modular files)**
   - Created `items/consumables.ts` - Healing/buff potions (5 items)
   - Created `items/quest-items.ts` - Quest-specific items (13 items)
   - Created `items/weapons.ts` - Swords, bows, axes (7 items)
   - Created `items/armor.ts` - Defensive equipment (6 items)
   - Created `items/accessories.ts` - Rings, amulets (6 items)
   - Created `items/food.ts` - Cooked meals (3 items)
   - Updated ExpandedItems.ts to re-export from modules

**Benefits:**
- Improved code organization and maintainability
- Easier to find and edit specific content types
- Better separation of concerns
- Maintains backward compatibility (all exports unchanged)

**Verification:**
- ✅ TypeCheck: Only 7 pre-existing errors (no new errors)
- ✅ Lint: Only 25 pre-existing warnings in tests
- ✅ Build: Compiles successfully

Addresses PR #67 review comments from Sourcery AI bot.
