+++ src/consts/DB_SEED/ExpandedItems.ts
@@ -1,543 +1,27 @@
-import { BUFF_TYPES } from './BuffTypes';
-import { ITEM_TYPE } from './ItemTypes';
-
 /**
  * Expanded Item Database
  * Comprehensive collection of items for the game expansion
  * Includes weapons, armor, consumables, and quest items
  */
 
-// Keep the original 4 items
+// Re-export the original 4 items
 export { DB_SEED_ITEMS } from './Items';
 
-// New Consumables (Healing & Buffs)
-export const CONSUMABLE_ITEMS = [
-	{
-		id: 5,
-		name: 'Greater Health Potion',
-		type: ITEM_TYPE.USABLE,
-		buffType: 0,
-		description: 'A potent healing potion that recovers 10 Health Points.',
-		script: 'rec hp 10;',
-		texture: 'red_potion',
-		sfx: 'heal',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 6,
-		name: 'Super Health Potion',
-		type: ITEM_TYPE.USABLE,
-		buffType: 0,
-		description: 'An extremely powerful potion that recovers 25 Health Points.',
-		script: 'rec hp 25;',
-		texture: 'red_potion',
-		sfx: 'heal',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 7,
-		name: 'Defense Potion',
-		type: ITEM_TYPE.USABLE,
-		buffType: BUFF_TYPES.DEF01,
-		description: 'Hardens your skin like stone. Increases DEFENSE by 5 for 60 seconds.',
-		script: 'buff def 5 60;',
-		texture: 'atk_potion',
-		sfx: 'heal',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 8,
-		name: 'Speed Potion',
-		type: ITEM_TYPE.USABLE,
-		buffType: BUFF_TYPES.SPD01,
-		description: 'Makes you light as a feather. Increases SPEED by 10 for 60 seconds.',
-		script: 'buff spd 10 60;',
-		texture: 'atk_potion',
-		sfx: 'heal',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 9,
-		name: 'Elixir of Power',
-		type: ITEM_TYPE.USABLE,
-		buffType: BUFF_TYPES.ATK01,
-		description: 'Ultimate power in a bottle. Increases ATTACK by 15 for 120 seconds.',
-		script: 'buff atk 15 120;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-];
+// Import and re-export all item categories
+export { CONSUMABLE_ITEMS } from './items/consumables';
+export { QUEST_ITEMS } from './items/quest-items';
+export { WEAPON_ITEMS } from './items/weapons';
+export { ARMOR_ITEMS } from './items/armor';
+export { ACCESSORY_ITEMS } from './items/accessories';
+export { FOOD_ITEMS } from './items/food';
 
-// Quest Items
-export const QUEST_ITEMS = [
-	{
-		id: 10,
-		name: 'Ancient Fragment',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'A mysterious fragment pulsing with ancient power. Required for a quest.',
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'item_pickup',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 11,
-		name: 'Dark Essence',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'Shadowy energy extracted from dark creatures. Sought by the Sage.',
-		script: '',
-		texture: 'atk_potion',
-		sfx: 'item_pickup',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 25,
-		name: 'Frost Seal Fragment',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'One of five fragments needed to seal the Shadow Lord.',
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'item_pickup',
-		stackable: false,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 26,
-		name: 'Flame Seal Fragment',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'One of five fragments needed to seal the Shadow Lord.',
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'item_pickup',
-		stackable: false,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 50,
-		name: "Lumberjack's Axe",
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: "The lumberjack's prized axe. Should return it to him.",
-		script: '',
-		texture: 'mighty_sword',
-		sfx: 'item_pickup',
-		stackable: false,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 51,
-		name: 'Glowing Mushroom',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'A luminescent mushroom from the cave. The alchemist needs these.',
-		script: '',
-		texture: 'red_potion',
-		sfx: 'item_pickup',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 53,
-		name: "Baker's Ring",
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'A golden ring belonging to the town baker.',
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'item_pickup',
-		stackable: false,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 54,
-		name: "Farmer's Hat",
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: "A well-worn farmer's hat lost in the fields.",
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'item_pickup',
-		stackable: false,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 55,
-		name: "Blacksmith's Hammer",
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: "A master blacksmith's favorite hammer.",
-		script: '',
-		texture: 'mighty_sword',
-		sfx: 'item_pickup',
-		stackable: false,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 85,
-		name: 'Ancient Artifact',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'A rare artifact from ancient times. Collectors pay well for these.',
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'item_pickup',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 91,
-		name: 'Wild Berries',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'Fresh berries found in the wild. Great for cooking.',
-		script: '',
-		texture: 'red_potion',
-		sfx: 'item_pickup',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 92,
-		name: 'Fresh Meat',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'Freshly hunted meat from wild beasts.',
-		script: '',
-		texture: 'red_potion',
-		sfx: 'item_pickup',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 100,
-		name: 'Ancient Seal',
-		type: ITEM_TYPE.QUEST,
-		buffType: 0,
-		description: 'The complete Ancient Seal, assembled from five fragments. Can seal ultimate evil.',
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'item_pickup',
-		stackable: false,
-		inventoryScale: 1.7,
-	},
-];
-
-// Weapons
-export const WEAPON_ITEMS = [
-	{
-		id: 15,
-		name: 'Legendary Blade',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'A legendary sword said to have slain dragons. Permanently increases ATTACK by 10.',
-		script: 'perm atk 10;',
-		texture: 'mighty_sword',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'weapon',
-	},
-	{
-		id: 60,
-		name: "Hunter's Bow",
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'A finely crafted bow. Permanently increases ATTACK by 6.',
-		script: 'perm atk 6;',
-		texture: 'mighty_sword',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'weapon',
-	},
-	{
-		id: 61,
-		name: 'Iron Sword',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'A standard iron sword. Permanently increases ATTACK by 3.',
-		script: 'perm atk 3;',
-		texture: 'mighty_sword',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'weapon',
-	},
-	{
-		id: 62,
-		name: 'Steel Axe',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'A heavy battle axe. Permanently increases ATTACK by 5.',
-		script: 'perm atk 5;',
-		texture: 'mighty_sword',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'weapon',
-	},
-	{
-		id: 63,
-		name: 'Mage Staff',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'A staff imbued with magical energy. Permanently increases ATTACK by 7.',
-		script: 'perm atk 7;',
-		texture: 'mighty_sword',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'weapon',
-	},
-	{
-		id: 32,
-		name: 'Frost Blade',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'A blade forged from eternal ice. Permanently increases ATTACK by 8 and adds ice damage.',
-		script: 'perm atk 8;',
-		texture: 'mighty_sword',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'weapon',
-		elementalDamage: 'ice',
-	},
-	{
-		id: 99,
-		name: 'Sword of Light',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'The ultimate weapon, blessed by the gods. Permanently increases ATTACK by 20.',
-		script: 'perm atk 20;',
-		texture: 'mighty_sword',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'weapon',
-	},
-];
-
-// Armor
-export const ARMOR_ITEMS = [
-	{
-		id: 20,
-		name: 'Epic Armor',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.DEF01,
-		description: 'Masterwork armor forged by legendary blacksmiths. Permanently increases DEFENSE by 8.',
-		script: 'perm def 8;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'armor',
-	},
-	{
-		id: 30,
-		name: 'Fire Resistance Armor',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.DEF01,
-		description:
-			'Armor enchanted with fire protection. Permanently increases DEFENSE by 6 and reduces fire damage.',
-		script: 'perm def 6;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'armor',
-		resistance: { fire: 0.5 },
-	},
-	{
-		id: 31,
-		name: 'Ice Resistance Armor',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.DEF01,
-		description:
-			'Armor enchanted with frost protection. Permanently increases DEFENSE by 6 and reduces ice damage.',
-		script: 'perm def 6;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'armor',
-		resistance: { ice: 0.5 },
-	},
-	{
-		id: 64,
-		name: 'Leather Armor',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.DEF01,
-		description: 'Basic leather protection. Permanently increases DEFENSE by 2.',
-		script: 'perm def 2;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'armor',
-	},
-	{
-		id: 65,
-		name: 'Chainmail',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.DEF01,
-		description: 'Strong metal links provide good protection. Permanently increases DEFENSE by 4.',
-		script: 'perm def 4;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'armor',
-	},
-	{
-		id: 66,
-		name: 'Plate Armor',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.DEF01,
-		description: 'Heavy plate armor for maximum protection. Permanently increases DEFENSE by 6.',
-		script: 'perm def 6;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'armor',
-	},
-];
-
-// Accessories
-export const ACCESSORY_ITEMS = [
-	{
-		id: 70,
-		name: 'Shadow Cloak',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.SPD01,
-		description: 'A cloak woven from shadows. Permanently increases SPEED by 5 and evasion.',
-		script: 'perm spd 5;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'accessory',
-	},
-	{
-		id: 80,
-		name: 'Dungeon Explorer Badge',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: 0,
-		description: 'Badge of an elite dungeon explorer. Increases treasure drop rates.',
-		script: '',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'accessory',
-		bonusDropRate: 1.25, // 25% more drops
-	},
-	{
-		id: 90,
-		name: 'Speed Boots',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.SPD01,
-		description: 'Enchanted boots that make you incredibly fast. Permanently increases SPEED by 10.',
-		script: 'perm spd 10;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'accessory',
-	},
-	{
-		id: 67,
-		name: 'Ring of Power',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.ATK02,
-		description: 'A ring humming with power. Permanently increases ATTACK by 4.',
-		script: 'perm atk 4;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'accessory',
-	},
-	{
-		id: 68,
-		name: 'Ring of Defense',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: BUFF_TYPES.DEF01,
-		description: 'A ring that hardens the skin. Permanently increases DEFENSE by 3.',
-		script: 'perm def 3;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'accessory',
-	},
-	{
-		id: 69,
-		name: 'Amulet of Health',
-		type: ITEM_TYPE.EQUIPMENT,
-		buffType: 0,
-		description: 'An amulet that boosts vitality. Permanently increases MAX HP by 20.',
-		script: 'perm hp 20;',
-		texture: 'treasure_chest',
-		sfx: 'equip_item',
-		stackable: false,
-		inventoryScale: 1.7,
-		equipSlot: 'accessory',
-	},
-];
-
-// Food Items (Cooked from ingredients)
-export const FOOD_ITEMS = [
-	{
-		id: 52,
-		name: 'Alchemist Potion',
-		type: ITEM_TYPE.USABLE,
-		buffType: BUFF_TYPES.ATK01,
-		description: 'A special potion brewed by the alchemist. Increases ATTACK by 8 for 90 seconds.',
-		script: 'buff atk 8 90;',
-		texture: 'atk_potion',
-		sfx: 'heal',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 93,
-		name: 'Cooked Meal',
-		type: ITEM_TYPE.USABLE,
-		buffType: 0,
-		description: 'A hearty cooked meal. Recovers 5 HP and grants temporary buffs.',
-		script: 'rec hp 5; buff atk 3 60;',
-		texture: 'red_potion',
-		sfx: 'heal',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-	{
-		id: 94,
-		name: 'Hero Feast',
-		type: ITEM_TYPE.USABLE,
-		buffType: 0,
-		description: 'A feast fit for heroes. Recovers 15 HP and grants major buffs.',
-		script: 'rec hp 15; buff atk 7 120; buff def 5 120;',
-		texture: 'treasure_chest',
-		sfx: 'heal',
-		stackable: true,
-		inventoryScale: 1.7,
-	},
-];
+// Import for combining
+import { CONSUMABLE_ITEMS } from './items/consumables';
+import { QUEST_ITEMS } from './items/quest-items';
+import { WEAPON_ITEMS } from './items/weapons';
+import { ARMOR_ITEMS } from './items/armor';
+import { ACCESSORY_ITEMS } from './items/accessories';
+import { FOOD_ITEMS } from './items/food';
 
 // Combine all expanded items
 export const ALL_EXPANDED_ITEMS = [
