+++ src/consts/DB_SEED/Skills.ts
@@ -1,4 +1,7 @@
 import { ISkill } from '../../types/SkillTypes';
+import { WARRIOR_SKILLS } from './skills/warrior-skills';
+import { MAGE_SKILLS } from './skills/mage-skills';
+import { ROGUE_SKILLS } from './skills/rogue-skills';
 
 /**
  * SKILL TREES
@@ -8,458 +11,8 @@ import { ISkill } from '../../types/SkillTypes';
  * - Rogue: Fast, evasive, critical hit specialist
  */
 
-// ===== WARRIOR SKILL TREE =====
-export const WARRIOR_SKILLS: ISkill[] = [
-	// Tier 1
-	{
-		id: 1,
-		name: 'Strength Training',
-		description: 'Basic combat training. +2 ATTACK.',
-		tree: 'warrior',
-		tier: 1,
-		icon: 'skill_str',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'atk', value: 2 }],
-	},
-	{
-		id: 2,
-		name: 'Toughness',
-		description: 'Harden your body. +3 DEFENSE.',
-		tree: 'warrior',
-		tier: 1,
-		icon: 'skill_def',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'def', value: 3 }],
-	},
-	{
-		id: 3,
-		name: 'Vitality',
-		description: 'Increase your life force. +10 MAX HP.',
-		tree: 'warrior',
-		tier: 1,
-		icon: 'skill_hp',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'hp', value: 10 }],
-	},
-
-	// Tier 2
-	{
-		id: 4,
-		name: 'Power Strike',
-		description: 'Devastating attack that deals 150% damage.',
-		tree: 'warrior',
-		tier: 2,
-		icon: 'skill_power',
-		maxRank: 3,
-		prerequisites: [1],
-		cost: 2,
-		effects: [{ type: 'active', active: 'power_strike' }],
-	},
-	{
-		id: 5,
-		name: 'Iron Skin',
-		description: 'Passive: Reduce all damage taken by 10%.',
-		tree: 'warrior',
-		tier: 2,
-		icon: 'skill_iron',
-		maxRank: 3,
-		prerequisites: [2],
-		cost: 2,
-		effects: [{ type: 'passive', passive: 'iron_skin' }],
-	},
-	{
-		id: 6,
-		name: 'Battle Fury',
-		description: '+5 ATTACK, -2 DEFENSE. High risk, high reward.',
-		tree: 'warrior',
-		tier: 2,
-		icon: 'skill_fury',
-		maxRank: 1,
-		prerequisites: [1],
-		cost: 2,
-		effects: [
-			{ type: 'stat', stat: 'atk', value: 5 },
-			{ type: 'stat', stat: 'def', value: -2 },
-		],
-	},
-
-	// Tier 3
-	{
-		id: 7,
-		name: 'Whirlwind',
-		description: 'Spin attack hitting all nearby enemies.',
-		tree: 'warrior',
-		tier: 3,
-		icon: 'skill_whirl',
-		maxRank: 2,
-		prerequisites: [4],
-		cost: 3,
-		effects: [{ type: 'active', active: 'whirlwind' }],
-	},
-	{
-		id: 8,
-		name: 'Last Stand',
-		description: 'Passive: Cannot be reduced below 1 HP once per battle.',
-		tree: 'warrior',
-		tier: 3,
-		icon: 'skill_last',
-		maxRank: 1,
-		prerequisites: [3, 5],
-		cost: 3,
-		effects: [{ type: 'passive', passive: 'last_stand' }],
-	},
-
-	// Tier 4
-	{
-		id: 9,
-		name: 'Titan Slam',
-		description: 'Massive attack that stuns enemies. 200% damage.',
-		tree: 'warrior',
-		tier: 4,
-		icon: 'skill_slam',
-		maxRank: 2,
-		prerequisites: [7],
-		cost: 4,
-		effects: [{ type: 'active', active: 'titan_slam' }],
-	},
-	{
-		id: 10,
-		name: 'Unstoppable',
-		description: 'Passive: Immune to stun, freeze, and knockback.',
-		tree: 'warrior',
-		tier: 4,
-		icon: 'skill_unstop',
-		maxRank: 1,
-		prerequisites: [8],
-		cost: 4,
-		effects: [{ type: 'passive', passive: 'unstoppable' }],
-	},
-
-	// Tier 5 - Ultimate
-	{
-		id: 11,
-		name: 'Berserker Rage',
-		description: 'Ultimate: Enter berserk mode. +100% ATTACK for 20 seconds.',
-		tree: 'warrior',
-		tier: 5,
-		icon: 'skill_berserk',
-		maxRank: 1,
-		prerequisites: [9, 10],
-		cost: 5,
-		effects: [{ type: 'active', active: 'berserker_rage' }],
-	},
-];
-
-// ===== MAGE SKILL TREE =====
-export const MAGE_SKILLS: ISkill[] = [
-	// Tier 1
-	{
-		id: 101,
-		name: 'Arcane Knowledge',
-		description: 'Study the arcane arts. +3 ATTACK (magic power).',
-		tree: 'mage',
-		tier: 1,
-		icon: 'skill_arcane',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'atk', value: 3 }],
-	},
-	{
-		id: 102,
-		name: 'Mana Pool',
-		description: 'Expand your mana reserves. +15 MAX HP (mana).',
-		tree: 'mage',
-		tier: 1,
-		icon: 'skill_mana',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'hp', value: 15 }],
-	},
-	{
-		id: 103,
-		name: 'Spell Focus',
-		description: 'Improve spell accuracy. +2 HIT.',
-		tree: 'mage',
-		tier: 1,
-		icon: 'skill_focus',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'hit', value: 2 }],
-	},
-
-	// Tier 2
-	{
-		id: 104,
-		name: 'Fireball',
-		description: 'Hurl a ball of flame at enemies.',
-		tree: 'mage',
-		tier: 2,
-		icon: 'skill_fire',
-		maxRank: 3,
-		prerequisites: [101],
-		cost: 2,
-		effects: [{ type: 'active', active: 'fireball' }],
-	},
-	{
-		id: 105,
-		name: 'Ice Shard',
-		description: 'Launch shards of ice that slow enemies.',
-		tree: 'mage',
-		tier: 2,
-		icon: 'skill_ice',
-		maxRank: 3,
-		prerequisites: [101],
-		cost: 2,
-		effects: [{ type: 'active', active: 'ice_shard' }],
-	},
-	{
-		id: 106,
-		name: 'Mana Shield',
-		description: 'Passive: Convert 20% damage to mana instead of HP.',
-		tree: 'mage',
-		tier: 2,
-		icon: 'skill_shield',
-		maxRank: 1,
-		prerequisites: [102],
-		cost: 2,
-		effects: [{ type: 'passive', passive: 'mana_shield' }],
-	},
-
-	// Tier 3
-	{
-		id: 107,
-		name: 'Chain Lightning',
-		description: 'Lightning that jumps between enemies.',
-		tree: 'mage',
-		tier: 3,
-		icon: 'skill_lightning',
-		maxRank: 2,
-		prerequisites: [104],
-		cost: 3,
-		effects: [{ type: 'active', active: 'chain_lightning' }],
-	},
-	{
-		id: 108,
-		name: 'Frost Nova',
-		description: 'Freeze all nearby enemies in place.',
-		tree: 'mage',
-		tier: 3,
-		icon: 'skill_nova',
-		maxRank: 2,
-		prerequisites: [105],
-		cost: 3,
-		effects: [{ type: 'active', active: 'frost_nova' }],
-	},
-	{
-		id: 109,
-		name: 'Arcane Mastery',
-		description: 'Passive: All spells cost 25% less mana and cooldown faster.',
-		tree: 'mage',
-		tier: 3,
-		icon: 'skill_mastery',
-		maxRank: 1,
-		prerequisites: [103, 106],
-		cost: 3,
-		effects: [{ type: 'passive', passive: 'arcane_mastery' }],
-	},
-
-	// Tier 4
-	{
-		id: 110,
-		name: 'Meteor Storm',
-		description: 'Rain meteors from the sky. Massive AoE damage.',
-		tree: 'mage',
-		tier: 4,
-		icon: 'skill_meteor',
-		maxRank: 2,
-		prerequisites: [107],
-		cost: 4,
-		effects: [{ type: 'active', active: 'meteor_storm' }],
-	},
-	{
-		id: 111,
-		name: 'Teleport',
-		description: 'Instantly teleport a short distance. Dodge attacks.',
-		tree: 'mage',
-		tier: 4,
-		icon: 'skill_teleport',
-		maxRank: 1,
-		prerequisites: [109],
-		cost: 4,
-		effects: [{ type: 'active', active: 'teleport' }],
-	},
-
-	// Tier 5 - Ultimate
-	{
-		id: 112,
-		name: 'Apocalypse',
-		description: 'Ultimate: Channel pure destruction. Obliterate everything.',
-		tree: 'mage',
-		tier: 5,
-		icon: 'skill_apocalypse',
-		maxRank: 1,
-		prerequisites: [110, 111],
-		cost: 5,
-		effects: [{ type: 'active', active: 'apocalypse' }],
-	},
-];
-
-// ===== ROGUE SKILL TREE =====
-export const ROGUE_SKILLS: ISkill[] = [
-	// Tier 1
-	{
-		id: 201,
-		name: 'Agility',
-		description: 'Move with grace. +5 SPEED.',
-		tree: 'rogue',
-		tier: 1,
-		icon: 'skill_agility',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'spd', value: 5 }],
-	},
-	{
-		id: 202,
-		name: 'Precision',
-		description: 'Strike with accuracy. +3 HIT.',
-		tree: 'rogue',
-		tier: 1,
-		icon: 'skill_precision',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'hit', value: 3 }],
-	},
-	{
-		id: 203,
-		name: 'Evasion',
-		description: 'Dodge attacks easier. +3 FLEE.',
-		tree: 'rogue',
-		tier: 1,
-		icon: 'skill_evasion',
-		maxRank: 5,
-		cost: 1,
-		effects: [{ type: 'stat', stat: 'flee', value: 3 }],
-	},
-
-	// Tier 2
-	{
-		id: 204,
-		name: 'Critical Strike',
-		description: 'Passive: +15% critical hit chance.',
-		tree: 'rogue',
-		tier: 2,
-		icon: 'skill_crit',
-		maxRank: 3,
-		prerequisites: [202],
-		cost: 2,
-		effects: [{ type: 'passive', passive: 'critical_strike' }],
-	},
-	{
-		id: 205,
-		name: 'Backstab',
-		description: 'Attack from behind for 200% damage.',
-		tree: 'rogue',
-		tier: 2,
-		icon: 'skill_backstab',
-		maxRank: 3,
-		prerequisites: [201],
-		cost: 2,
-		effects: [{ type: 'active', active: 'backstab' }],
-	},
-	{
-		id: 206,
-		name: 'Shadow Step',
-		description: 'Dash behind enemy. Great for positioning.',
-		tree: 'rogue',
-		tier: 2,
-		icon: 'skill_dash',
-		maxRank: 2,
-		prerequisites: [203],
-		cost: 2,
-		effects: [{ type: 'active', active: 'shadow_step' }],
-	},
-
-	// Tier 3
-	{
-		id: 207,
-		name: 'Poison Blade',
-		description: 'Attacks apply poison damage over time.',
-		tree: 'rogue',
-		tier: 3,
-		icon: 'skill_poison',
-		maxRank: 2,
-		prerequisites: [204],
-		cost: 3,
-		effects: [{ type: 'passive', passive: 'poison_blade' }],
-	},
-	{
-		id: 208,
-		name: 'Smoke Bomb',
-		description: 'Become invisible for 5 seconds. Escape danger.',
-		tree: 'rogue',
-		tier: 3,
-		icon: 'skill_smoke',
-		maxRank: 2,
-		prerequisites: [206],
-		cost: 3,
-		effects: [{ type: 'active', active: 'smoke_bomb' }],
-	},
-	{
-		id: 209,
-		name: 'Deadly Precision',
-		description: 'Passive: Critical hits deal 250% damage instead of 200%.',
-		tree: 'rogue',
-		tier: 3,
-		icon: 'skill_deadly',
-		maxRank: 1,
-		prerequisites: [204, 205],
-		cost: 3,
-		effects: [{ type: 'passive', passive: 'deadly_precision' }],
-	},
-
-	// Tier 4
-	{
-		id: 210,
-		name: 'Assassinate',
-		description: 'Execute low-health enemies instantly (below 20% HP).',
-		tree: 'rogue',
-		tier: 4,
-		icon: 'skill_assassinate',
-		maxRank: 2,
-		prerequisites: [207, 209],
-		cost: 4,
-		effects: [{ type: 'active', active: 'assassinate' }],
-	},
-	{
-		id: 211,
-		name: 'Vanish',
-		description: 'Disappear completely. Reappear with guaranteed critical.',
-		tree: 'rogue',
-		tier: 4,
-		icon: 'skill_vanish',
-		maxRank: 1,
-		prerequisites: [208],
-		cost: 4,
-		effects: [{ type: 'active', active: 'vanish' }],
-	},
-
-	// Tier 5 - Ultimate
-	{
-		id: 212,
-		name: 'Death Mark',
-		description: 'Ultimate: Mark enemy for death. Next attack deals 500% damage.',
-		tree: 'rogue',
-		tier: 5,
-		icon: 'skill_death',
-		maxRank: 1,
-		prerequisites: [210, 211],
-		cost: 5,
-		effects: [{ type: 'active', active: 'death_mark' }],
-	},
-];
+// Re-export skill arrays
+export { WARRIOR_SKILLS, MAGE_SKILLS, ROGUE_SKILLS };
 
 // Combine all skills
 export const ALL_SKILLS: ISkill[] = [...WARRIOR_SKILLS, ...MAGE_SKILLS, ...ROGUE_SKILLS];
