+++ src/consts/DB_SEED/items/quest-items.ts
@@ -1,0 +1,164 @@
+import { ITEM_TYPE } from '../ItemTypes';
+
+/**
+ * Quest Items
+ * Special items required for completing quests
+ */
+export const QUEST_ITEMS = [
+	{
+		id: 10,
+		name: 'Ancient Fragment',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'A mysterious fragment pulsing with ancient power. Required for a quest.',
+		script: '',
+		texture: 'treasure_chest',
+		sfx: 'item_pickup',
+		stackable: true,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 11,
+		name: 'Dark Essence',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'Shadowy energy extracted from dark creatures. Sought by the Sage.',
+		script: '',
+		texture: 'atk_potion',
+		sfx: 'item_pickup',
+		stackable: true,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 25,
+		name: 'Frost Seal Fragment',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'One of five fragments needed to seal the Shadow Lord.',
+		script: '',
+		texture: 'treasure_chest',
+		sfx: 'item_pickup',
+		stackable: false,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 26,
+		name: 'Flame Seal Fragment',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'One of five fragments needed to seal the Shadow Lord.',
+		script: '',
+		texture: 'treasure_chest',
+		sfx: 'item_pickup',
+		stackable: false,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 50,
+		name: "Lumberjack's Axe",
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: "The lumberjack's prized axe. Should return it to him.",
+		script: '',
+		texture: 'mighty_sword',
+		sfx: 'item_pickup',
+		stackable: false,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 51,
+		name: 'Glowing Mushroom',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'A luminescent mushroom from the cave. The alchemist needs these.',
+		script: '',
+		texture: 'red_potion',
+		sfx: 'item_pickup',
+		stackable: true,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 53,
+		name: "Baker's Ring",
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'A golden ring belonging to the town baker.',
+		script: '',
+		texture: 'treasure_chest',
+		sfx: 'item_pickup',
+		stackable: false,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 54,
+		name: "Farmer's Hat",
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: "A well-worn farmer's hat lost in the fields.",
+		script: '',
+		texture: 'treasure_chest',
+		sfx: 'item_pickup',
+		stackable: false,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 55,
+		name: "Blacksmith's Hammer",
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: "A master blacksmith's favorite hammer.",
+		script: '',
+		texture: 'mighty_sword',
+		sfx: 'item_pickup',
+		stackable: false,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 85,
+		name: 'Ancient Artifact',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'A rare artifact from ancient times. Collectors pay well for these.',
+		script: '',
+		texture: 'treasure_chest',
+		sfx: 'item_pickup',
+		stackable: true,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 91,
+		name: 'Wild Berries',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'Fresh berries found in the wild. Great for cooking.',
+		script: '',
+		texture: 'red_potion',
+		sfx: 'item_pickup',
+		stackable: true,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 92,
+		name: 'Fresh Meat',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'Freshly hunted meat from wild beasts.',
+		script: '',
+		texture: 'red_potion',
+		sfx: 'item_pickup',
+		stackable: true,
+		inventoryScale: 1.7,
+	},
+	{
+		id: 100,
+		name: 'Ancient Seal',
+		type: ITEM_TYPE.QUEST,
+		buffType: 0,
+		description: 'The complete Ancient Seal, assembled from five fragments. Can seal ultimate evil.',
+		script: '',
+		texture: 'treasure_chest',
+		sfx: 'item_pickup',
+		stackable: false,
+		inventoryScale: 1.7,
+	},
+];
