+++ src/consts/DB_SEED/items/weapons.ts
@@ -1,0 +1,101 @@
+import { BUFF_TYPES } from '../BuffTypes';
+import { ITEM_TYPE } from '../ItemTypes';
+
+/**
+ * Weapon Items
+ * Swords, bows, axes, staves, and other offensive equipment
+ */
+export const WEAPON_ITEMS = [
+	{
+		id: 15,
+		name: 'Legendary Blade',
+		type: ITEM_TYPE.EQUIP,
+		buffType: BUFF_TYPES.ATK02,
+		description: 'A legendary sword said to have slain dragons. Permanently increases ATTACK by 10.',
+		script: 'perm atk 10;',
+		texture: 'mighty_sword',
+		sfx: 'equip_item',
+		stackable: false,
+		inventoryScale: 1.7,
+		equipSlot: 'weapon',
+	},
+	{
+		id: 60,
+		name: "Hunter's Bow",
+		type: ITEM_TYPE.EQUIP,
+		buffType: BUFF_TYPES.ATK02,
+		description: 'A finely crafted bow. Permanently increases ATTACK by 6.',
+		script: 'perm atk 6;',
+		texture: 'mighty_sword',
+		sfx: 'equip_item',
+		stackable: false,
+		inventoryScale: 1.7,
+		equipSlot: 'weapon',
+	},
+	{
+		id: 61,
+		name: 'Iron Sword',
+		type: ITEM_TYPE.EQUIP,
+		buffType: BUFF_TYPES.ATK02,
+		description: 'A standard iron sword. Permanently increases ATTACK by 3.',
+		script: 'perm atk 3;',
+		texture: 'mighty_sword',
+		sfx: 'equip_item',
+		stackable: false,
+		inventoryScale: 1.7,
+		equipSlot: 'weapon',
+	},
+	{
+		id: 62,
+		name: 'Steel Axe',
+		type: ITEM_TYPE.EQUIP,
+		buffType: BUFF_TYPES.ATK02,
+		description: 'A heavy battle axe. Permanently increases ATTACK by 5.',
+		script: 'perm atk 5;',
+		texture: 'mighty_sword',
+		sfx: 'equip_item',
+		stackable: false,
+		inventoryScale: 1.7,
+		equipSlot: 'weapon',
+	},
+	{
+		id: 63,
+		name: 'Mage Staff',
+		type: ITEM_TYPE.EQUIP,
+		buffType: BUFF_TYPES.ATK02,
+		description: 'A staff imbued with magical energy. Permanently increases ATTACK by 7.',
+		script: 'perm atk 7;',
+		texture: 'mighty_sword',
+		sfx: 'equip_item',
+		stackable: false,
+		inventoryScale: 1.7,
+		equipSlot: 'weapon',
+	},
+	{
+		id: 32,
+		name: 'Frost Blade',
+		type: ITEM_TYPE.EQUIP,
+		buffType: BUFF_TYPES.ATK02,
+		description: 'A blade forged from eternal ice. Permanently increases ATTACK by 8 and adds ice damage.',
+		script: 'perm atk 8;',
+		texture: 'mighty_sword',
+		sfx: 'equip_item',
+		stackable: false,
+		inventoryScale: 1.7,
+		equipSlot: 'weapon',
+		elementalDamage: 'ice',
+	},
+	{
+		id: 99,
+		name: 'Sword of Light',
+		type: ITEM_TYPE.EQUIP,
+		buffType: BUFF_TYPES.ATK02,
+		description: 'The ultimate weapon, blessed by the gods. Permanently increases ATTACK by 20.',
+		script: 'perm atk 20;',
+		texture: 'mighty_sword',
+		sfx: 'equip_item',
+		stackable: false,
+		inventoryScale: 1.7,
+		equipSlot: 'weapon',
+	},
+];
