+++ src/consts/DB_SEED/skills/mage-skills.ts
@@ -1,0 +1,159 @@
+import { ISkill } from '../../../types/SkillTypes';
+
+/**
+ * MAGE SKILL TREE
+ * Magic user with elemental spells
+ * Fire, ice, lightning, and arcane magic
+ */
+export const MAGE_SKILLS: ISkill[] = [
+	// Tier 1
+	{
+		id: 101,
+		name: 'Arcane Knowledge',
+		description: 'Study the arcane arts. +3 ATTACK (magic power).',
+		tree: 'mage',
+		tier: 1,
+		icon: 'skill_arcane',
+		maxRank: 5,
+		cost: 1,
+		effects: [{ type: 'stat', stat: 'atk', value: 3 }],
+	},
+	{
+		id: 102,
+		name: 'Mana Pool',
+		description: 'Expand your mana reserves. +15 MAX HP (mana).',
+		tree: 'mage',
+		tier: 1,
+		icon: 'skill_mana',
+		maxRank: 5,
+		cost: 1,
+		effects: [{ type: 'stat', stat: 'hp', value: 15 }],
+	},
+	{
+		id: 103,
+		name: 'Spell Focus',
+		description: 'Improve spell accuracy. +2 HIT.',
+		tree: 'mage',
+		tier: 1,
+		icon: 'skill_focus',
+		maxRank: 5,
+		cost: 1,
+		effects: [{ type: 'stat', stat: 'hit', value: 2 }],
+	},
+
+	// Tier 2
+	{
+		id: 104,
+		name: 'Fireball',
+		description: 'Hurl a ball of flame at enemies.',
+		tree: 'mage',
+		tier: 2,
+		icon: 'skill_fire',
+		maxRank: 3,
+		prerequisites: [101],
+		cost: 2,
+		effects: [{ type: 'active', active: 'fireball' }],
+	},
+	{
+		id: 105,
+		name: 'Ice Shard',
+		description: 'Launch shards of ice that slow enemies.',
+		tree: 'mage',
+		tier: 2,
+		icon: 'skill_ice',
+		maxRank: 3,
+		prerequisites: [101],
+		cost: 2,
+		effects: [{ type: 'active', active: 'ice_shard' }],
+	},
+	{
+		id: 106,
+		name: 'Mana Shield',
+		description: 'Passive: Convert 20% damage to mana instead of HP.',
+		tree: 'mage',
+		tier: 2,
+		icon: 'skill_shield',
+		maxRank: 1,
+		prerequisites: [102],
+		cost: 2,
+		effects: [{ type: 'passive', passive: 'mana_shield' }],
+	},
+
+	// Tier 3
+	{
+		id: 107,
+		name: 'Chain Lightning',
+		description: 'Lightning that jumps between enemies.',
+		tree: 'mage',
+		tier: 3,
+		icon: 'skill_lightning',
+		maxRank: 2,
+		prerequisites: [104],
+		cost: 3,
+		effects: [{ type: 'active', active: 'chain_lightning' }],
+	},
+	{
+		id: 108,
+		name: 'Frost Nova',
+		description: 'Freeze all nearby enemies in place.',
+		tree: 'mage',
+		tier: 3,
+		icon: 'skill_nova',
+		maxRank: 2,
+		prerequisites: [105],
+		cost: 3,
+		effects: [{ type: 'active', active: 'frost_nova' }],
+	},
+	{
+		id: 109,
+		name: 'Arcane Mastery',
+		description: 'Passive: All spells cost 25% less mana and cooldown faster.',
+		tree: 'mage',
+		tier: 3,
+		icon: 'skill_mastery',
+		maxRank: 1,
+		prerequisites: [103, 106],
+		cost: 3,
+		effects: [{ type: 'passive', passive: 'arcane_mastery' }],
+	},
+
+	// Tier 4
+	{
+		id: 110,
+		name: 'Meteor Storm',
+		description: 'Rain meteors from the sky. Massive AoE damage.',
+		tree: 'mage',
+		tier: 4,
+		icon: 'skill_meteor',
+		maxRank: 2,
+		prerequisites: [107],
+		cost: 4,
+		effects: [{ type: 'active', active: 'meteor_storm' }],
+	},
+	{
+		id: 111,
+		name: 'Teleport',
+		description: 'Instantly teleport a short distance. Dodge attacks.',
+		tree: 'mage',
+		tier: 4,
+		icon: 'skill_teleport',
+		maxRank: 1,
+		prerequisites: [109],
+		cost: 4,
+		effects: [{ type: 'active', active: 'teleport' }],
+	},
+
+	// Tier 5 - Ultimate
+	{
+		id: 112,
+		name: 'Apocalypse',
+		description: 'Ultimate: Channel pure destruction. Obliterate everything.',
+		tree: 'mage',
+		tier: 5,
+		icon: 'skill_apocalypse',
+		maxRank: 1,
+		prerequisites: [110, 111],
+		cost: 5,
+		effects: [{ type: 'active', active: 'apocalypse' }],
+	},
+];
