+++ src/consts/DB_SEED/skills/rogue-skills.ts
@@ -1,0 +1,159 @@
+import { ISkill } from '../../../types/SkillTypes';
+
+/**
+ * ROGUE SKILL TREE
+ * Fast, evasive, critical hit specialist
+ * Stealth, poison, and deadly precision
+ */
+export const ROGUE_SKILLS: ISkill[] = [
+	// Tier 1
+	{
+		id: 201,
+		name: 'Agility',
+		description: 'Move with grace. +5 SPEED.',
+		tree: 'rogue',
+		tier: 1,
+		icon: 'skill_agility',
+		maxRank: 5,
+		cost: 1,
+		effects: [{ type: 'stat', stat: 'spd', value: 5 }],
+	},
+	{
+		id: 202,
+		name: 'Precision',
+		description: 'Strike with accuracy. +3 HIT.',
+		tree: 'rogue',
+		tier: 1,
+		icon: 'skill_precision',
+		maxRank: 5,
+		cost: 1,
+		effects: [{ type: 'stat', stat: 'hit', value: 3 }],
+	},
+	{
+		id: 203,
+		name: 'Evasion',
+		description: 'Dodge attacks easier. +3 FLEE.',
+		tree: 'rogue',
+		tier: 1,
+		icon: 'skill_evasion',
+		maxRank: 5,
+		cost: 1,
+		effects: [{ type: 'stat', stat: 'flee', value: 3 }],
+	},
+
+	// Tier 2
+	{
+		id: 204,
+		name: 'Critical Strike',
+		description: 'Passive: +15% critical hit chance.',
+		tree: 'rogue',
+		tier: 2,
+		icon: 'skill_crit',
+		maxRank: 3,
+		prerequisites: [202],
+		cost: 2,
+		effects: [{ type: 'passive', passive: 'critical_strike' }],
+	},
+	{
+		id: 205,
+		name: 'Backstab',
+		description: 'Attack from behind for 200% damage.',
+		tree: 'rogue',
+		tier: 2,
+		icon: 'skill_backstab',
+		maxRank: 3,
+		prerequisites: [201],
+		cost: 2,
+		effects: [{ type: 'active', active: 'backstab' }],
+	},
+	{
+		id: 206,
+		name: 'Shadow Step',
+		description: 'Dash behind enemy. Great for positioning.',
+		tree: 'rogue',
+		tier: 2,
+		icon: 'skill_dash',
+		maxRank: 2,
+		prerequisites: [203],
+		cost: 2,
+		effects: [{ type: 'active', active: 'shadow_step' }],
+	},
+
+	// Tier 3
+	{
+		id: 207,
+		name: 'Poison Blade',
+		description: 'Attacks apply poison damage over time.',
+		tree: 'rogue',
+		tier: 3,
+		icon: 'skill_poison',
+		maxRank: 2,
+		prerequisites: [204],
+		cost: 3,
+		effects: [{ type: 'passive', passive: 'poison_blade' }],
+	},
+	{
+		id: 208,
+		name: 'Smoke Bomb',
+		description: 'Become invisible for 5 seconds. Escape danger.',
+		tree: 'rogue',
+		tier: 3,
+		icon: 'skill_smoke',
+		maxRank: 2,
+		prerequisites: [206],
+		cost: 3,
+		effects: [{ type: 'active', active: 'smoke_bomb' }],
+	},
+	{
+		id: 209,
+		name: 'Deadly Precision',
+		description: 'Passive: Critical hits deal 250% damage instead of 200%.',
+		tree: 'rogue',
+		tier: 3,
+		icon: 'skill_deadly',
+		maxRank: 1,
+		prerequisites: [204, 205],
+		cost: 3,
+		effects: [{ type: 'passive', passive: 'deadly_precision' }],
+	},
+
+	// Tier 4
+	{
+		id: 210,
+		name: 'Assassinate',
+		description: 'Execute low-health enemies instantly (below 20% HP).',
+		tree: 'rogue',
+		tier: 4,
+		icon: 'skill_assassinate',
+		maxRank: 2,
+		prerequisites: [207, 209],
+		cost: 4,
+		effects: [{ type: 'active', active: 'assassinate' }],
+	},
+	{
+		id: 211,
+		name: 'Vanish',
+		description: 'Disappear completely. Reappear with guaranteed critical.',
+		tree: 'rogue',
+		tier: 4,
+		icon: 'skill_vanish',
+		maxRank: 1,
+		prerequisites: [208],
+		cost: 4,
+		effects: [{ type: 'active', active: 'vanish' }],
+	},
+
+	// Tier 5 - Ultimate
+	{
+		id: 212,
+		name: 'Death Mark',
+		description: 'Ultimate: Mark enemy for death. Next attack deals 500% damage.',
+		tree: 'rogue',
+		tier: 5,
+		icon: 'skill_death',
+		maxRank: 1,
+		prerequisites: [210, 211],
+		cost: 5,
+		effects: [{ type: 'active', active: 'death_mark' }],
+	},
+];
